+++ FireworkFlutterSDKSample/lib/screens/more/more_screen.dart
@@ -8,7 +8,7 @@ import 'package:fw_flutter_sdk_example/utils/host_app_service.dart';
 import '../../generated/l10n.dart';
 import '../../widgets/fw_app_bar.dart';
 
-const fwNativeVersionOfAndroid = '6.24.1';
+const fwNativeVersionOfAndroid = '6.25.0';
 
 class MoreScreen extends StatefulWidget {
   const MoreScreen({
@@ -24,6 +24,8 @@ class _MoreScreenState extends State<MoreScreen> {
   DataTrackingLevel _dataTrackingLevel = DataTrackingLevel.all;
   LivestreamPlayerDesignVersion _livestreamPlayerVersion =
       LivestreamPlayerDesignVersion.v1;
+  ShortVideoPlayerDesignVersion _shortVideoPlayerVersion =
+      ShortVideoPlayerDesignVersion.v1;
   @override
   void initState() {
     super.initState();
@@ -46,6 +48,11 @@ class _MoreScreenState extends State<MoreScreen> {
       HostAppService.getInstance().getLivestreamPlayerVersion().then((value) {
         setState(() {
           _livestreamPlayerVersion = value ?? LivestreamPlayerDesignVersion.v1;
+        });
+      });
+      HostAppService.getInstance().getShortVideoPlayerVersion().then((value) {
+        setState(() {
+          _shortVideoPlayerVersion = value ?? ShortVideoPlayerDesignVersion.v1;
         });
       });
     });
@@ -228,6 +235,53 @@ class _MoreScreenState extends State<MoreScreen> {
                       ),
                       actions: [
                         for (var w in livestreamPlayerDesignVersionWidgetArray)
+                          w,
+                      ],
+                      cancelButton: CupertinoActionSheetAction(
+                        child: Text(
+                          S.of(context).cancel,
+                        ),
+                        onPressed: () {
+                          Navigator.of(context).pop();
+                        },
+                      ),
+                    );
+                  },
+                );
+              },
+            ),
+            _buildItem(
+              context: context,
+              title: S
+                  .of(context)
+                  .changeShortVideoPlayerVersion(_shortVideoPlayerVersion.name),
+              onTap: () {
+                showCupertinoModalPopup(
+                  context: context,
+                  builder: (context) {
+                    final shortVideoPlayerDesignVersionWidgetArray =
+                        ShortVideoPlayerDesignVersion.values.map(
+                      (version) => CupertinoActionSheetAction(
+                        isDefaultAction: true,
+                        onPressed: () {
+                          HostAppService.getInstance()
+                              .cacheShortVideoPlayerVersion(version);
+                          setState(() {
+                            _shortVideoPlayerVersion = version;
+                          });
+                          Navigator.of(context).pop();
+                        },
+                        child: Text(
+                          version.name,
+                        ),
+                      ),
+                    );
+                    return CupertinoActionSheet(
+                      title: Text(
+                        S.of(context).selectShortVideoPlayerVersion,
+                      ),
+                      actions: [
+                        for (var w in shortVideoPlayerDesignVersionWidgetArray)
                           w,
                       ],
                       cancelButton: CupertinoActionSheetAction(